added icons to feed and began optimizing flatlist with initial number of items to render
